+++ sigils-of-outri/Zombie.h
@@ -8,7 +8,7 @@ public:
     Zombie(const char * name);
     ~Zombie();
     
-    virtual int activate(Stats & stats, Character & target) override;
+    virtual float activate(Stats & stats, Character & target) override;
 
 private:
